Datebase Project
Statement 
Design and implement a relational database for The University Accommodation Office (descriptions of the system can be found at the end of the document). Build a Java interface to the database suitable for the database application. 
 
Tasks 
This includes setting up a database schema and providing access methods to this in the form of queries and views. The steps for doing this include: 
Creating a conceptual schema in the form of an ER diagram.   
    Deriving a relational schema from the ER diagram.  
    Validating the relations using normalization technique.  
    Implementing this schema using SQL.  
    Populating the database with a set of typical data. 
    The data should be significant but manageable.  
    Defining SQL queries which could be used in the Java interface.  
    Design and implementation of the Java interface, using JDBC to connect the database.